+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/product/create/CreateProductViewModel.groovy
@@ -1,9 +1,10 @@
 package life.qbic.portal.offermanager.components.product.create
 
+
 import groovy.beans.Bindable
 import life.qbic.datamodel.dtos.business.ProductCategory
+import life.qbic.datamodel.dtos.business.facilities.Facility
 import life.qbic.datamodel.dtos.business.services.ProductUnit
-
 
 /**
  * <h1>Holds all values that the user specifies in the CreateProductView</h1>
@@ -17,11 +18,15 @@ class CreateProductViewModel {
     @Bindable Boolean productNameValid
     @Bindable String productDescription
     @Bindable Boolean productDescriptionValid
-    @Bindable String productUnitPrice //todo make sure to cast the string to a double in the use case
-    @Bindable Boolean productUnitPriceValid
+    @Bindable String internalUnitPrice
+    @Bindable String externalUnitPrice
+    @Bindable Boolean internalUnitPriceValid
+    @Bindable Boolean externalUnitPriceValid
     @Bindable ProductUnit productUnit
     @Bindable Boolean productUnitValid
     @Bindable ProductCategory productCategory
     @Bindable Boolean productCategoryValid
+    @Bindable Facility productFacility
+    @Bindable Boolean productFacilityValid
 
 }